requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pyyaml==6.0.1
numpy==1.26.2
sentence-transformers==2.7.0
scikit-learn==1.3.2
